fix: Linking policies to other objects
Fixes an issue where it wasn't possible to create a link to a policy
from an object of a different type outside of linking targettables.
Currently creating an edge to a policy from an object would silently
fail since the policy is not in the topology at the time of creating the
links.

Signed-off-by: Michael Nairn <mnairn@redhat.com>